Refactor reinstall button
